2 steps forward 1 step back
optimized sections of code, but also didn't fix the STARTING issue with cards like i thought so last push XD

+++ Assets/Scripts/Slime/SO_SlimePart.cs
@@ -31,6 +31,6 @@ public class SO_SlimePart : ScriptableObject
     public CardComponentType CardComponentType;
     public string PartName;
     public Sprite ImgToDisplay;
-    public float Power;
+    public int Power;
     public int Cost;
 }
